eval/grader/requirements.txt
# Pinned versions for the LOCKED grader. Do NOT bump without a GRADER_HASH
# rev — prior scoreboard entries become incomparable on any dep change.
cadquery-ocp==7.9.3.1
trimesh==4.11.5
numpy==2.4.4
